Fixed Various Things

+++ src/persistence/dao/AfferisceDao.java
@@ -7,7 +7,7 @@ import model.Afferisce;
 public interface AfferisceDao {
 	
 	public void save(Afferisce afferisce);  // Create
-	public Afferisce findByPrimaryKey(int id);     // Retrieve
+	public Afferisce findByPrimaryKey(long id);     // Retrieve
 	public List<Afferisce> findAll();       
 	public void update(Afferisce afferisce); //Update
 	public void delete(Afferisce afferisce); //Delete
